+++ android/src/main/kotlin/dev/ltag/stone_payments/StonePaymentsPlugin.kt
@@ -18,8 +18,10 @@ import io.flutter.plugin.common.MethodChannel.Result as Res
 /** StonePaymentsPlugin */
 class StonePaymentsPlugin : FlutterPlugin, MethodCallHandler, Activity() {
     private lateinit var channel: MethodChannel
-    var transactionObject = TransactionObject()
     var context: Context = this;
+    var transactionObject = TransactionObject()
+    var paymentUsecase: PaymentUsecase? = null
+    var printerUsecase: PrinterUsecase? = null
 
     companion object {
         var flutterBinaryMessenger: BinaryMessenger? = null
@@ -30,13 +32,13 @@ class StonePaymentsPlugin : FlutterPlugin, MethodCallHandler, Activity() {
         flutterBinaryMessenger = flutterPluginBinding.binaryMessenger;
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, "stone_payments")
         channel.setMethodCallHandler(this)
+        // Inicialize as propriedades aqui
+        paymentUsecase = PaymentUsecase(this)
+        printerUsecase = PrinterUsecase(this)
     }
 
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Res) {
         val activateUsecase: ActivateUsecase? = ActivateUsecase(context)
-        val paymentUsecase: PaymentUsecase? = PaymentUsecase(this,)
-        val printerUsecase: PrinterUsecase? = PrinterUsecase(this,)
-
         when (call.method) {
             "activateStone" -> {
                 try {
@@ -127,18 +129,65 @@ class StonePaymentsPlugin : FlutterPlugin, MethodCallHandler, Activity() {
                     result.error("UNAVAILABLE", "Cannot Activate", e.toString())
                 }
             }
-//          "cancel-payment" -> {
-//              try {
-//                  payment!!.cancel { resp ->
-//                      when (resp) {
-//                          is Result.Success<*> -> result.success(resp.data.toString())
-//                          else -> result.error("Error", resp.toString(), resp.toString())
-//                      }
-//                  }
-//              } catch (e: Exception) {
-//                  result.error("UNAVAILABLE", "Cannot cancel", e.toString())
-//              }
-//          }
+            "abortPayment" -> {
+                try {
+                    paymentUsecase!!.doAbort() { resp ->
+                        when (resp) {
+                            is Result.Success<String> -> result.success(
+                                resp.data
+                            )
+                            else -> result.error("Error", resp.toString(), resp.toString())
+                        }
+                    }
+                } catch (e: Exception) {
+                    result.error("UNAVAILABLE", "Cannot Activate", e.toString())
+                }
+            }
+            "cancelPayment" -> {
+                try {
+                    paymentUsecase!!.doCancelWithITK(
+                       call.argument("initiatorTransactionKey")!!,
+                       call.argument("printReceipt"),
+                   ) { resp ->
+                        when (resp) {
+                            is Result.Success<*> -> result.success(resp.data.toString())
+                            else -> result.error("Error", resp.toString(), resp.toString())
+                        }
+                    }
+                } catch (e: Exception) {
+                    result.error("UNAVAILABLE", "Cannot cancel", e.toString())
+                }
+            }
+            // "cancelPaymentWithATK" -> {
+            //     try {
+            //         paymentUsecase!!.doCancelWithITK(
+            //            call.argument("acquirerTransactionKey")!!,
+            //            call.argument("printReceipt"),
+            //        ) { resp ->
+            //             when (resp) {
+            //                 is Result.Success<*> -> result.success(resp.data.toString())
+            //                 else -> result.error("Error", resp.toString(), resp.toString())
+            //             }
+            //         }
+            //     } catch (e: Exception) {
+            //         result.error("UNAVAILABLE", "Cannot cancel", e.toString())
+            //     }
+            // }
+            "cancelPaymentWithAuthorizationCode" -> {
+                try {
+                    paymentUsecase!!.doCancelWithAuthorizationCode(
+                       call.argument("authorizationCode")!!,
+                       call.argument("printReceipt"),
+                   ) { resp ->
+                        when (resp) {
+                            is Result.Success<*> -> result.success(resp.data.toString())
+                            else -> result.error("Error", resp.toString(), resp.toString())
+                        }
+                    }
+                } catch (e: Exception) {
+                    result.error("UNAVAILABLE", "Cannot cancel", e.toString())
+                }
+            }
             else -> {
                 result.notImplemented()
             }