+++ android/src/main/kotlin/dev/ltag/stone_payments/usecases/PaymentUsecase.kt
@@ -6,6 +6,7 @@ import android.util.Log
 import com.google.gson.Gson
 import br.com.stone.posandroid.providers.PosPrintReceiptProvider
 import br.com.stone.posandroid.providers.PosTransactionProvider
+
 import dev.ltag.stone_payments.Result
 import dev.ltag.stone_payments.StonePaymentsPlugin
 import dev.ltag.stone_payments.core.Helper
@@ -20,12 +21,15 @@ import stone.application.enums.TypeOfTransactionEnum
 import stone.application.interfaces.StoneActionCallback
 import stone.application.interfaces.StoneCallbackInterface
 import stone.database.transaction.TransactionObject
+import stone.database.transaction.TransactionDAO
+import stone.providers.CancellationProvider;
 import stone.utils.Stone
 
 class PaymentUsecase(
     private val stonePayments: StonePaymentsPlugin,
 ) {
     private val context = stonePayments.context
+    lateinit var provider: PosTransactionProvider
 
     fun doPayment(
         value: Double,
@@ -47,7 +51,7 @@ class PaymentUsecase(
             val newValue: Int = (value * 100).toInt()
             transactionObject.amount = newValue.toString()
 
-            val provider = PosTransactionProvider(
+            provider = PosTransactionProvider(
                 context,
                 transactionObject,
                 Stone.getUserModel(0),
@@ -129,7 +133,6 @@ class PaymentUsecase(
             Log.d("ERROR", e.toString())
             callback(Result.Error(e))
         }
-
     }
 
     fun doTransaction(
@@ -144,15 +147,14 @@ class PaymentUsecase(
 
             val transactionObject = stonePayments.transactionObject
 
-            transactionObject.instalmentTransaction =
-                InstalmentTransactionEnum.getAt(installment - 1)
+            transactionObject.instalmentTransaction = InstalmentTransactionEnum.getAt(installment - 1)
             transactionObject.typeOfTransaction = TypeOfTransactionEnum.values()[type]
 
             transactionObject.isCapture = true
             val newValue: Int = (value * 100).toInt()
             transactionObject.amount = newValue.toString()
 
-            val provider = PosTransactionProvider(
+            provider = PosTransactionProvider(
                 context,
                 transactionObject,
                 Stone.getUserModel(0),
@@ -233,7 +235,134 @@ class PaymentUsecase(
             Log.d("ERROR", e.toString())
             callback(Result.Error(e))
         }
+    }
 
+    fun doAbort(callback: (Result<String>) -> Unit) {
+        try {
+            val transactionObject = stonePayments.transactionObject
+
+            if(provider == null){
+                callback(Result.Error(Exception("PROVIDER NOT FOUND")))
+                return
+            }
+
+            provider.abortPayment()
+            callback(Result.Success("ABORTED"))
+        } catch (e: Exception) {
+            Log.d("ERROR", e.toString())
+            callback(Result.Error(e))
+        }
+    }
+
+    fun doCancelWithITK(initiatorTransactionKey: String, print: Boolean?, callback: (Result<String>) -> Unit) {
+        try {
+            val transactionDAO = TransactionDAO(context)
+            val transactionObject = transactionDAO.findTransactionWithInitiatorTransactionKey(initiatorTransactionKey)
+
+            if (transactionObject == null) {
+                callback(Result.Error(Exception("TRANSACTION NOT FOUND")))
+                return
+            }
+
+            return cancel(transactionObject, print, callback)
+
+        } catch (e: Exception) {
+            Log.d("ERROR", e.toString())
+            callback(Result.Error(e));
+        }
+    }
+
+    // fun doCancelWithATK(acquirerTransactionKey: String, print: Boolean?, callback: (Result<String>) -> Unit) {
+    //     try {
+    //         val transactionDAO = TransactionDAO(context)
+    //         val transactionObject = transactionDAO.findTransactionWithATK(acquirerTransactionKey)
+
+    //         if (transactionObject == null) {
+    //             callback(Result.Error(Exception("TRANSACTION NOT FOUND")))
+    //             return
+    //         }
+
+    //         return cancel(transactionObject, print, callback)
+
+    //     } catch (e: Exception) {
+    //         Log.d("ERROR", e.toString())
+    //         callback(Result.Error(e));
+    //     }
+    // }
+
+    fun doCancelWithAuthorizationCode(authorizationCode: String, print: Boolean?, callback: (Result<String>) -> Unit) {
+        try {
+            val transactionDAO = TransactionDAO(context)
+            val transactionObject = transactionDAO.findTransactionWithAuthorizationCode(authorizationCode)
+
+            if (transactionObject == null) {
+                callback(Result.Error(Exception("TRANSACTION NOT FOUND")))
+                return
+            }
+
+            return cancel(transactionObject, print, callback)
+
+        } catch (e: Exception) {
+            Log.d("ERROR", e.toString())
+            callback(Result.Error(e));
+        }
+    }
+
+    private fun cancel(transactionObject: TransactionObject, print: Boolean?, callback: (Result<String>) -> Unit,) {
+        if(transactionObject == null) {
+            callback(Result.Error(Exception("NOT FOUND")))
+            return
+        }
+
+        val provider = CancellationProvider(
+            context,
+            transactionObject,
+        )
+
+        provider.setConnectionCallback(object : StoneCallbackInterface {
+
+            override fun onSuccess() {
+                if(print == true) {
+                    val posPrintReceiptProvider =
+                        PosPrintReceiptProvider(
+                            context, transactionObject,
+                            ReceiptType.MERCHANT,
+                        );
+
+                    posPrintReceiptProvider.connectionCallback = object :
+                        StoneCallbackInterface {
+
+                        override fun onSuccess() {
+
+                            Log.d("SUCCESS", transactionObject.toString())
+                            
+                        }
+
+                        override fun onError() {
+                            Log.d("ERRORPRINT", transactionObject.toString())
+
+                        }
+                    }
+
+                    posPrintReceiptProvider.execute()
+                }
+
+                val serializableTransaction = SerializableTransactionObject.from(transactionObject)
+
+                val jsonString = Gson().toJson(serializableTransaction)
+                callback(Result.Success(jsonString))
+
+            }
+
+            override fun onError() {
+
+                Log.d("RESULT", "ERROR")
+
+                callback(Result.Error(Exception("ERROR")));
+            }
+        })
+
+        provider.execute()
     }
 
     private fun sendAMessage(message: String) {